+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - 更新文档徽标链接 | Update docs badge link
 - 更换文档发布状态徽标 | Change docs publish status badge
 - 升级 VitePress 版本 | Upgrade VitePress version
+- 添加事件介绍 | Add event introduction